requirements.txt
Django==2.2.6
djangorestframework==3.10.3
pynamodb==4.1.0
